Remade
From the query command if multiqc flag is used, run the run_multiqc function by giving it the abolute path of the query result.txt + user specified directory to save the multiqc in. If html flag is specified, automatically open the html doc

If csv flag is specified, run the create_csv function. Input is the result of the query which returns all the metrics from a given set of samples

Use print_query function to simply print query