More flexible querying
Now the user doesn't have to worry about what joins are needed for certain filters or selects. They can put any combination of select and/or filter and it should produce a meaningful result.

This also extends to --multiqc or --csv flags, the user doesn't have to worry about selects, they can just add the filters they want and that's it.

-o output option is no longer required, as this interfers with stdout option. Instead, the user is prompted if -o is missing when using -multiqc/csv flag

Please try queries which may break it.