Save command - better error handling
Save will now prevent duplicate batch entries

Prevent mis-matched metadata/multiqc JSONs from being saved

-d and -s are no longer required, which allows -bm and -cm to be added independently. A check ensures when using -d you must also use -s

Note: If a batch is sucessfully saved, it will print out to stout, and this means it has actually been loaded into the database. This means if a bulk save has been performed and 1 of the entries has a fatal error (i.e. mis-match or duplicate entry), every entry prior to that one will be in the database, and every entry after and including the failed entry will not be in the database. If you were to then try re-save the batch, you will have to remove the entries already saved into the database to prevent the duplicate entry exception from raising - (note, I'm not sure if this is what we want or not, this can be changed by deleting one line of code so not a big deal).