Update save.py
If an exception is raised during bulk save - all entries are rolled back. However once everything is saved it will be committed.

This means if then the -bm command encounters a exception, the saved multiqc/metadata will not be rolled back, etc. only the -bm will be rolled back. Same for -cm